# services/feed-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Feed Service API
  version: 1.0.0
  description: |
    RSS/Atom feed management and fetching microservice for News MCP platform.

    ## Features
    - Feed CRUD operations with UUID-based identification
    - Automatic feed fetching via scheduler
    - Feed quality and health monitoring
    - Content scraping (httpx/playwright)
    - Auto-analysis pipeline (categorization, sentiment, OSINT)
    - Event-driven integration via RabbitMQ
    - Bulk operations support

    ## Authentication
    All endpoints (except /health and /) require JWT authentication.
    Include token in Authorization header: `Bearer <token>`

    ## Event Publishing
    The service publishes events to RabbitMQ:
    - `feed.created` - New feed added
    - `feed.updated` - Feed configuration changed
    - `feed.deleted` - Feed removed
    - `feed.fetched` - Feed fetch completed
    - `feed.fetch_failed` - Feed fetch failed

  contact:
    name: News MCP Platform
    url: https://github.com/news-mcp
  license:
    name: MIT
servers:
  - url: http://localhost:8101
    description: Development server
  - url: http://localhost:8101/api/v1
    description: API v1 endpoints
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/feeds:
    get:
      summary: List feeds
      description: |
        Get list of all feeds with pagination and advanced filtering.

        Supports filtering by:
        - Active status
        - Feed status (ACTIVE, PAUSED, ERROR, INACTIVE)
        - Category
        - Health score range
      tags:
        - Feeds
      parameters:
        - name: skip
          in: query
          description: Number of records to skip (pagination offset)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          description: Maximum number of records to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: status
          in: query
          description: Filter by feed status
          schema:
            type: string
            enum: [ACTIVE, PAUSED, ERROR, INACTIVE]
        - name: category
          in: query
          description: Filter by category name
          schema:
            type: string
            example: technology
        - name: health_score_min
          in: query
          description: Minimum health score (0-100)
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: health_score_max
          in: query
          description: Maximum health score (0-100)
          schema:
            type: integer
            minimum: 0
            maximum: 100
      responses:
        '200':
          description: List of feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feed'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create feed
      description: |
        Add a new RSS/Atom feed to the system.

        After creation:
        - Initial health record is created
        - Feed is automatically fetched in background
        - `feed.created` event is published to RabbitMQ
      tags:
        - Feeds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedCreate'
            examples:
              basic_feed:
                summary: Basic RSS feed
                value:
                  name: TechCrunch
                  url: https://techcrunch.com/feed/
                  fetch_interval: 60
              full_content_scraping:
                summary: Feed with content scraping
                value:
                  name: Hacker News
                  url: https://news.ycombinator.com/rss
                  fetch_interval: 30
                  scrape_full_content: true
                  scrape_method: httpx
                  categories: ["technology", "startup"]
              auto_analysis:
                summary: Feed with auto-analysis enabled
                value:
                  name: Financial Times
                  url: https://www.ft.com/rss/home
                  fetch_interval: 15
                  enable_categorization: true
                  enable_finance_sentiment: true
                  categories: ["finance", "business"]
      responses:
        '201':
          description: Feed created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "fetch_interval must be between 5 and 1440 minutes"
        '409':
          description: Feed with this URL already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Feed with this URL already exists"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/v1/feeds/{feed_id}/fetch:
    post:
      summary: Manually trigger feed fetch
      description: |
        Trigger immediate feed fetch and parsing.

        The fetch operation runs asynchronously in background.
        Use GET /api/v1/feeds/{feed_id} to check last_fetched_at timestamp.

        Publishes `feed.fetched` or `feed.fetch_failed` event.
      tags:
        - Feeds
      parameters:
        - $ref: '#/components/parameters/FeedId'
      responses:
        '200':
          description: Fetch triggered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Fetch triggered for feed 123e4567-e89b-12d3-a456-426614174000"
                  feed_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: Feed is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Feed is not active"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/v1/feeds/bulk-fetch:
    post:
      summary: Bulk fetch feeds
      description: |
        Trigger fetch for multiple feeds or all active feeds.

        If `feed_ids` is empty or null, all active feeds will be fetched.
        Use `force: true` to fetch even recently fetched feeds (ignore rate limiting).

        All fetches are performed asynchronously in background.
      tags:
        - Bulk Operations
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkFetchRequest'
            examples:
              fetch_specific:
                summary: Fetch specific feeds
                value:
                  feed_ids:
                    - "123e4567-e89b-12d3-a456-426614174000"
                    - "223e4567-e89b-12d3-a456-426614174001"
                  force: false
              fetch_all:
                summary: Fetch all active feeds
                value:
                  feed_ids: []
                  force: false
              force_fetch:
                summary: Force fetch all feeds
                value:
                  feed_ids: []
                  force: true
      responses:
        '200':
          description: Bulk fetch scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkFetchResponse'
              example:
                total_feeds: 15
                successful_fetches: 0
                failed_fetches: 0
                total_new_items: 0
                details:
                  - feed_id: "123e4567-e89b-12d3-a456-426614174000"
                    feed_name: "TechCrunch"
                    status: "scheduled"
                  - feed_id: "223e4567-e89b-12d3-a456-426614174001"
                    feed_name: "Hacker News"
                    status: "scheduled"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Auth Service.

        Token structure:
        ```json
        {
          "sub": "user_id",
          "exp": 1234567890,
          "iat": 1234567890
        }
        ```
  parameters:
    FeedId:
      name: feed_id
      in: path
      required: true
      description: Feed UUID
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
...
  schemas:
    Feed:
      type: object
      required:
        - id
        - name
        - url
        - fetch_interval
        - is_active
        - status
        - health_score
        - consecutive_failures
        - total_items
        - items_last_24h
        - scrape_full_content
        - scrape_method
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Feed UUID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Feed display name
          example: "TechCrunch"
        url:
          type: string
          format: uri
          description: RSS/Atom feed URL
          example: "https://techcrunch.com/feed/"
        description:
          type: string
          nullable: true
          description: Feed description
          example: "Latest technology news and startup coverage"
        fetch_interval:
          type: integer
          minimum: 5
          maximum: 1440
          description: Fetch interval in minutes (5 min to 24 hours)
          example: 60
        is_active:
          type: boolean
          description: Whether feed is active and should be fetched
          example: true
        status:
          type: string
          enum: [ACTIVE, PAUSED, ERROR, INACTIVE]
          description: Feed status
          example: "ACTIVE"
        last_fetched_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of last successful fetch
          example: "2025-10-15T14:30:00Z"
        health_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Feed health score (0-100)
          example: 95
        consecutive_failures:
          type: integer
          minimum: 0
          description: Number of consecutive fetch failures
          example: 0
        total_items:
          type: integer
          minimum: 0
          description: Total number of items in feed
          example: 1234
        items_last_24h:
          type: integer
          minimum: 0
          description: Number of new items in last 24 hours
          example: 15
        scrape_full_content:
          type: boolean
          description: Whether to scrape full article content
          example: false
        scrape_method:
          type: string
          enum: [auto, httpx, playwright]
          description: Scraping method (auto-detect, HTTP client, or headless browser)
          example: "auto"
        enable_categorization:
          type: boolean
          description: Enable automatic article categorization (Stage 1)
          example: false
        enable_finance_sentiment:
          type: boolean
          description: Enable finance-specific sentiment analysis (Stage 2)
          example: false
        enable_geopolitical_sentiment:
          type: boolean
          description: Enable geopolitical sentiment analysis (Stage 2)
          example: false
        enable_osint_analysis:
          type: boolean
          description: Enable OSINT event analysis (Stage 3)
          example: false
        created_at:
          type: string
          format: date-time
          description: Feed creation timestamp
          example: "2025-10-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Feed last update timestamp
          example: "2025-10-15T14:30:00Z"
        categories:
          type: array
          items:
            type: string
          description: Feed categories for organization
          example: ["technology", "startup", "ai"]
        legacy_id:
          type: integer
          nullable: true
          description: Legacy integer ID for backward compatibility
          example: null
    FeedCreate:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Feed display name
          example: "TechCrunch"
        url:
          type: string
          format: uri
          description: RSS/Atom feed URL
          example: "https://techcrunch.com/feed/"
        description:
          type: string
          nullable: true
          description: Feed description
          example: "Latest technology news and startup coverage"
        fetch_interval:
          type: integer
          minimum: 5
          maximum: 1440
          default: 60
          description: Fetch interval in minutes
          example: 60
        scrape_full_content:
          type: boolean
          default: false
          description: Whether to scrape full article content
          example: false
        scrape_method:
          type: string
          enum: [auto, httpx, playwright]
          default: "auto"
          description: Scraping method
          example: "auto"
        enable_categorization:
          type: boolean
          default: false
          description: Enable automatic categorization
          example: false
        enable_finance_sentiment:
          type: boolean
          default: false
          description: Enable finance sentiment analysis
          example: false
        enable_geopolitical_sentiment:
          type: boolean
          default: false
          description: Enable geopolitical sentiment analysis
          example: false
        enable_osint_analysis:
          type: boolean
          default: false
          description: Enable OSINT analysis
          example: false
        categories:
          type: array
          items:
            type: string
          default: []
          description: Initial feed categories
          example: ["technology", "startup"]
...
    BulkFetchRequest:
      type: object
      properties:
        feed_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
          description: Feed UUIDs to fetch (empty/null = all active feeds)
          example:
            - "123e4567-e89b-12d3-a456-426614174000"
            - "223e4567-e89b-12d3-a456-426614174001"
        force:
          type: boolean
          default: false
          description: Force fetch even if recently fetched
          example: false
    BulkFetchResponse:
      type: object
      required:
        - total_feeds
        - successful_fetches
        - failed_fetches
        - total_new_items
        - details
      properties:
        total_feeds:
          type: integer
          minimum: 0
          description: Total number of feeds scheduled
          example: 15
        successful_fetches:
          type: integer
          minimum: 0
          description: Number of successful fetches (updated by background tasks)
          example: 0
        failed_fetches:
          type: integer
          minimum: 0
          description: Number of failed fetches (updated by background tasks)
          example: 0
        total_new_items:
          type: integer
          minimum: 0
          description: Total new items fetched (updated by background tasks)
          example: 0
        details:
          type: array
          items:
            type: object
            properties:
              feed_id:
                type: string
                format: uuid
              feed_name:
                type: string
              status:
                type: string
                enum: [scheduled, success, failed]
          description: Per-feed details
          example:
            - feed_id: "123e4567-e89b-12d3-a456-426614174000"
              feed_name: "TechCrunch"
              status: "scheduled"
            - feed_id: "223e4567-e89b-12d3-a456-426614174001"
              feed_name: "Hacker News"
              status: "scheduled"
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Feed not found"
        type:
          type: string
          description: Error type/code
          example: "not_found"
  responses:
    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Could not validate credentials"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Feed 123e4567-e89b-12d3-a456-426614174000 not found"